app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile /home/gcc/work/projects/querido-diario/backend/app/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
backports-zoneinfo==0.2.1
    # via django
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
dj-database-url==1.2.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
django==4.1.7
    # via
    #   -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
    #   dj-database-url
    #   django-anymail
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-anymail[mailjet]==9.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
django-cors-headers==3.14.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
django-rest-passwordreset==1.3.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
djangorestframework==3.14.0
    # via
    #   -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
gunicorn==20.1.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
idna==3.4
    # via requests
kombu==5.2.4
    # via celery
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.5
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
python-decouple==3.8
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
pytz==2022.7.1
    # via
    #   celery
    #   djangorestframework
redis==4.5.1
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
requests==2.28.2
    # via
    #   -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
    #   django-anymail
six==1.16.0
    # via click-repl
sqlparse==0.4.3
    # via django
urllib3==1.26.14
    # via requests
validate-docbr==1.10.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise==6.4.0
    # via -r /home/gcc/work/projects/querido-diario/backend/app/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
